+++ packages/api/api/handlers/stellar_cartography.py
@@ -1,8 +1,12 @@
 """Shared turn-scoped Stellar Cartography sample route logic for Core and BFF callers."""
 
 from api.concepts.stellar_cartography.sample_at import sample_at as sample_at_turn
+from api.concepts.stellar_cartography.turn_summary import stellar_cartography_turn_summary
 from api.services.turn_concept_service import TurnConceptService
-from api.transport.concept_stellar_cartography import StellarCartographySampleResponse
+from api.transport.concept_stellar_cartography import (
+    StellarCartographySampleResponse,
+    StellarCartographyTurnSummaryResponse,
+)
 
 
 def sample_at(
@@ -17,3 +21,15 @@ def sample_at(
     turn = svc.get_turn_info(game_id, perspective, turn_number)
     payload = sample_at_turn(turn, x, y)
     return StellarCartographySampleResponse.model_validate(payload)
+
+
+def turn_summary(
+    svc: TurnConceptService,
+    game_id: int,
+    perspective: int,
+    turn_number: int,
+) -> StellarCartographyTurnSummaryResponse:
+    """Return lightweight cartography facts for the turn."""
+    turn = svc.get_turn_info(game_id, perspective, turn_number)
+    payload = stellar_cartography_turn_summary(turn)
+    return StellarCartographyTurnSummaryResponse.model_validate(payload)
