+++ packages/api/api/routers/game_concepts.py
@@ -2,11 +2,14 @@
 
 from fastapi import APIRouter, Depends, Query
 
-from api.handlers.stellar_cartography import sample_at
+from api.handlers.stellar_cartography import sample_at, turn_summary
 from api.handlers.warp_well import coordinate_in_well, warp_well_cells
 from api.services.deps import get_turn_concept_service
 from api.services.turn_concept_service import TurnConceptService
-from api.transport.concept_stellar_cartography import StellarCartographySampleResponse
+from api.transport.concept_stellar_cartography import (
+    StellarCartographySampleResponse,
+    StellarCartographyTurnSummaryResponse,
+)
 from api.transport.concept_warp_well import (
     CoordinateInWarpWellRequest,
     CoordinateInWarpWellResponse,
@@ -62,3 +65,17 @@ def get_stellar_cartography_sample(
 ) -> StellarCartographySampleResponse:
     """Return stacked Stellar Cartography tooltip entries at map cell ``(x, y)``."""
     return sample_at(svc, game_id, perspective, turn_number, x, y)
+
+
+@router.get(
+    "/{game_id}/{perspective}/turns/{turn_number}/concepts/stellar-cartography/summary",
+    response_model=StellarCartographyTurnSummaryResponse,
+)
+def get_stellar_cartography_turn_summary(
+    game_id: int,
+    perspective: int,
+    turn_number: int,
+    svc: TurnConceptService = Depends(get_turn_concept_service),
+) -> StellarCartographyTurnSummaryResponse:
+    """Return lightweight Stellar Cartography facts for the turn."""
+    return turn_summary(svc, game_id, perspective, turn_number)
