+++ packages/api/api/transport/concept_stellar_cartography.py
@@ -12,3 +12,8 @@ class StellarCartographySampleResponse(BaseModel):
     x: int
     y: int
     entries: list[StellarCartographySampleEntry] = Field(default_factory=list)
+
+
+class StellarCartographyTurnSummaryResponse(BaseModel):
+    ion_storm_count: int
+    nu_ion_storms: bool
