+++ packages/api/tests/test_game_concepts_router.py
@@ -116,3 +116,8 @@ class TestStellarCartographySample:
         assert data["y"] == 200
         layers = [e["layer"] for e in data["entries"]]
         assert "nebulae" in layers
+
+    def test_turn_summary_returns_ion_storm_availability(self, client):
+        resp = client.get("/v1/games/628580/1/turns/111/concepts/stellar-cartography/summary")
+        assert resp.status_code == 200
+        assert resp.json() == {"ion_storm_count": 3, "nu_ion_storms": True}
