+++ packages/bff/bff/core_client.py
@@ -7,7 +7,12 @@ from typing import TypeVar
 
 from api.diagnostics import NOOP_DIAGNOSTICS, Diagnostics
 from api.errors import NotFoundError, PlanetsConsoleError
-from api.handlers.stellar_cartography import sample_at as stellar_cartography_sample_at
+from api.handlers.stellar_cartography import (
+    sample_at as stellar_cartography_sample_at,
+)
+from api.handlers.stellar_cartography import (
+    turn_summary as stellar_cartography_turn_summary_handler,
+)
 from api.handlers.warp_well import coordinate_in_well, warp_well_cells
 from api.models.game import GameInfo, TurnInfo
 from api.planets_nu import PlanetsNuClient
@@ -19,7 +24,10 @@ from api.services.turn_concept_service import TurnConceptService
 from api.services.turn_load_service import TurnLoadService
 from api.storage import get_storage
 from api.storage.base import StorageBackend
-from api.transport.concept_stellar_cartography import StellarCartographySampleResponse
+from api.transport.concept_stellar_cartography import (
+    StellarCartographySampleResponse,
+    StellarCartographyTurnSummaryResponse,
+)
 from api.transport.concept_warp_well import (
     CoordinateInWarpWellRequest,
     CoordinateInWarpWellResponse,
@@ -199,6 +207,21 @@ class CoreClient:
                 turn_number,
                 x,
                 y,
+            )
+        )
+
+    def stellar_cartography_turn_summary(
+        self,
+        game_id: int,
+        perspective: int,
+        turn_number: int,
+    ) -> StellarCartographyTurnSummaryResponse:
+        return self._invoke(
+            lambda: stellar_cartography_turn_summary_handler(
+                self._concepts,
+                game_id,
+                perspective,
+                turn_number,
             )
         )
