+++ packages/frontend/src/components/MapGraph.tsx
@@ -1,33 +1,21 @@
 import {
-  type CSSProperties,
-  createContext,
   useCallback,
-  useContext,
   useEffect,
   useLayoutEffect,
   useMemo,
   useRef,
   useState,
 } from 'react'
-import { shallow } from 'zustand/shallow'
 import {
-  BaseEdge,
   ReactFlow,
-  Handle,
   Panel,
-  Position,
-  getStraightPath,
   useReactFlow,
   useStore,
   useStoreApi,
-  type Node,
-  type Edge,
-  type EdgeProps,
 } from '@xyflow/react'
 import '@xyflow/react/dist/style.css'
 import type { CombinedMapData, RouteMapWaypoint, AnalyticShellScope } from '../api/bff'
 import { StellarCartographyHoverPanel } from '../analytics/stellar-cartography/StellarCartographyHoverPanel'
-import { RasterFieldOverlay } from './RasterFieldOverlay'
 import type {
   CartographyLayerVisibility,
   StellarCartographySettingsGates,
@@ -36,33 +24,15 @@ import {
   filterWormholeEdgesForDisplayMode,
   type WormholeDisplayMode,
 } from '../analytics/stellar-cartography/wormholeDisplayMode'
-import {
-  buildStellarCartographyOverlayPaneShapes,
-  gameMapCellCenterToFlow,
-  type StellarCartographyOverlayRadialGradient,
-} from '../lib/stellarCartographyOverlay'
-import {
-  ionStormCloudPaneShapeToRasterField,
-  nebulaCloudPaneShapeToRasterField,
-} from '../lib/cartographyRasterFieldOverlay'
-import {
-  WORMHOLE_EDGE_OPACITY,
-  WORMHOLE_LINE_STROKE,
-  WORMHOLE_RECENTER_PULSE_MS,
-} from '../lib/stellarCartographyTheme'
-import { WormholeEndpointIconMark } from '../lib/wormholeEndpointIcon'
+import { WORMHOLE_RECENTER_PULSE_MS } from '../lib/stellarCartographyTheme'
 import {
   buildWormholeEndpointHoverIndex,
-  formatWormholeEndpointHoverLines,
-  type WormholeEndpointHoverInfo,
-  wormholeEndpointRecenterGameCoords,
   wormholeMapCellKey,
 } from '../lib/wormholeEndpointHover'
 import {
   buildPlanetSpatialGrid,
   findClosestPlanetWithinRadius,
   flowCenterToPlanet,
-  PLANET_CELL_CENTER_OFFSET,
   type PlanetSpatialGrid,
 } from '../lib/planetSpatialGrid'
 import { cn } from '../lib/utils'
@@ -76,520 +46,37 @@ import {
   buildWarpWellOverlayPaneLines,
   WARP_WELL_OVERLAY_ZOOM_THRESHOLD,
 } from '../lib/warpWellOverlay'
+import {
+  CELL_CENTER_OFFSET,
+  clientToFlowPosition,
+  flowCenterFromMapNode,
+  gameMapYToFlowCenterY,
+  safeZoomScale,
+} from './map-graph/geometry'
+import { nodeTypes, toFlowNodes } from './map-graph/nodes'
+import { edgeTypes, toEdges } from './map-graph/edges'
+import {
+  StellarCartographyOverlayPane,
+  collectWormholeEndpoints,
+} from './map-graph/StellarCartographyOverlayPane'
+import {
+  WORMHOLE_LINE_REVEAL_CLEAR_MS,
+  WormholeHoverContext,
+  WormholeLineRevealContext,
+  WormholeRecenterPulseContext,
+  type WormholeLineRevealApi,
+  type WormholeRecenterPulseTarget,
+} from './map-graph/stellarCartographyWormholeInteraction'
 
-const STELLAR_CARTOGRAPHY_NODE_PREFIX = 'stellar-cartography:'
-
-function collectWormholeEndpoints(
-  nodes: CombinedMapData['nodes'],
-  unknownEntrances: CombinedMapData['wormholeUnknownEntrances']
-): { x: number; y: number }[] {
-  const seen = new Set<string>()
-  const endpoints: { x: number; y: number }[] = []
-  const add = (x: number, y: number) => {
-    const key = `${x},${y}`
-    if (seen.has(key)) return
-    seen.add(key)
-    endpoints.push({ x, y })
-  }
-  for (const node of nodes) {
-    if (node.id.startsWith(STELLAR_CARTOGRAPHY_NODE_PREFIX)) {
-      add(Number(node.x), Number(node.y))
-    }
-  }
-  for (const entrance of unknownEntrances) {
-    add(entrance.x, entrance.y)
-  }
-  return endpoints
-}
-
-type MapNodeData = {
-  label?: string
-  ordinal: number
-  x: number
-  y: number
-  planet?: Record<string, unknown>
-  ownerName?: string | null
-}
-
-/** Stable node size in flow space so React Flow keeps node measurements through zoom. */
-const NODE_SIZE_FLOW = 12
 /** Fixed pixel size of the planet dot on screen (independent of zoom). */
 const DOT_PIXELS = 4
 /** On-screen size of a multi-hop route intermediate marker (smaller and quieter than planet dots). */
 const ROUTE_WAYPOINT_CROSS_PX = 5
 /** Mouse distance from dot center (px) at which the planet label is shown. */
 const PLANET_LABEL_HOVER_RADIUS_PX = 14
-/** Offset so node and edge targets use the center of the map cell (0.5, 0.5) as demarcated by grid lines at integers. */
-const CELL_CENTER_OFFSET = PLANET_CELL_CENTER_OFFSET
-
-/** Flow Y for React Flow (y grows downward); smaller game y sits lower on screen. */
-function gameMapYToFlowCenterY(py: number): number {
-  return -(py + CELL_CENTER_OFFSET)
-}
 
 /** Fraction of display area to leave as blank margin on each side when fitting initial view (0.1 = 10%). */
 const INITIAL_FIT_MARGIN = 0.1
-
-function safeZoomScale(scale: number | undefined): number {
-  return typeof scale === 'number' && Number.isFinite(scale) && scale > 0 ? scale : 1
-}
-
-/** Invisible handle at node center so edges connect to dot center. */
-const centerHandleStyle: CSSProperties = {
-  position: 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  opacity: 0,
-  width: 12,
-  height: 12,
-  minWidth: 12,
-  minHeight: 12,
-  border: 'none',
-  padding: 0,
-  background: 'transparent',
-  pointerEvents: 'none',
-}
-
-/** Invisible routing node; visible dot is drawn by the overlay. */
-function DotNode() {
-  return (
-    <div
-      className="relative"
-      style={{ width: NODE_SIZE_FLOW, height: NODE_SIZE_FLOW, minWidth: NODE_SIZE_FLOW, minHeight: NODE_SIZE_FLOW }}
-    >
-      <Handle type="target" position={Position.Left} id="t" style={centerHandleStyle} />
-      <Handle type="source" position={Position.Left} id="s" style={centerHandleStyle} />
-      {/* Planet labels are rendered in a separate overlay to keep pixel-stable positioning. */}
-    </div>
-  )
-}
-
-const nodeTypes = { dot: DotNode }
-
-const WormholeHoverContext = createContext<(lines: string[] | null) => void>(() => {})
-
-export type WormholeRecenterPulseTarget = {
-  mapX: number
-  mapY: number
-  token: number
-}
-
-const WormholeRecenterPulseContext = createContext<(mapX: number, mapY: number) => void>(() => {})
-
-const WORMHOLE_LINE_REVEAL_CLEAR_MS = 120
-
-type WormholeLineRevealApi = {
-  revealAt: (mapX: number, mapY: number) => void
-  scheduleClear: () => void
-  cancelClear: () => void
-}
-
-const WormholeLineRevealContext = createContext<WormholeLineRevealApi>({
-  revealAt: () => {},
-  scheduleClear: () => {},
-  cancelClear: () => {},
-})
-
-function recenterMapOnWormholeGameCell(
-  gameX: number,
-  gameY: number,
-  domNode: HTMLElement | null,
-  getViewport: () => { x: number; y: number; zoom: number },
-  setViewport: (vp: { x: number; y: number; zoom: number }) => void,
-  pulseAt: (mapX: number, mapY: number) => void
-): void {
-  const { cx, cy } = gameMapCellCenterToFlow(gameX, gameY)
-  recenterViewportOnFlowPoint(cx, cy, domNode, getViewport, setViewport)
-  pulseAt(gameX, gameY)
-}
-
-/** Custom edge keeps endpoints centered on dot nodes and stays visually 1px while zooming. */
-function StraightEdgeOnePixel(props: EdgeProps) {
-  const { sourceNode, targetNode, zoom } = useStore(
-    (s) => {
-      const nodeLookup = s.nodeLookup
-      return {
-        sourceNode: nodeLookup.get(props.source),
-        targetNode: nodeLookup.get(props.target),
-        zoom: s.transform[2],
-      }
-    },
-    shallow
-  )
-  const scale = safeZoomScale(zoom)
-  const half = NODE_SIZE_FLOW / 2
-  const sourceX = sourceNode ? sourceNode.position.x + half : props.sourceX
-  const sourceY = sourceNode ? sourceNode.position.y + half : props.sourceY
-  const targetX = targetNode ? targetNode.position.x + half : props.targetX
-  const targetY = targetNode ? targetNode.position.y + half : props.targetY
-  const [path] = getStraightPath({
-    sourceX,
-    sourceY,
-    targetX,
-    targetY,
-  })
-
-  const viaFlare =
-    props.data != null &&
-    typeof props.data === 'object' &&
-    (props.data as { viaFlare?: boolean }).viaFlare === true
-
-  return (
-    <BaseEdge
-      path={path}
-      style={{
-        stroke: viaFlare ? '#facc15' : '#b1b1b7',
-        strokeWidth: 1 / scale,
-        opacity: 0.5,
-        ...(viaFlare ? { strokeDasharray: `${4 / scale} ${3 / scale}` } : {}),
-      }}
-    />
-  )
-}
-
-const FLOW_POINT_EPS = 1e-6
-
-function dedupeConsecutiveFlowPoints(
-  points: { x: number; y: number }[]
-): { x: number; y: number }[] {
-  const out: { x: number; y: number }[] = []
-  for (const p of points) {
-    const last = out[out.length - 1]
-    if (last == null || Math.hypot(p.x - last.x, p.y - last.y) > FLOW_POINT_EPS) {
-      out.push(p)
-    }
-  }
-  return out
-}
-
-function buildSvgPathThroughFlowPoints(
-  points: { x: number; y: number }[]
-): string {
-  if (points.length < 2) return ''
-  let d = `M ${points[0].x} ${points[0].y}`
-  for (let i = 1; i < points.length; i += 1) {
-    d += ` L ${points[i].x} ${points[i].y}`
-  }
-  return d
-}
-
-type MapEdgeData = {
-  viaFlare?: boolean
-  waypointsInGame?: { x: number; y: number }[]
-}
-
-type WormholeEdgeData = {
-  isBidirectional?: boolean
-  sourceGameX?: number
-  sourceGameY?: number
-  targetGameX?: number
-  targetGameY?: number
-  wormholeExitOnly?: boolean
-}
-
-function wormholeHoverLabel(
-  data: WormholeEdgeData | undefined,
-  nearSource: boolean
-): string | null {
-  if (data == null) return null
-  const sx = data.sourceGameX
-  const sy = data.sourceGameY
-  const tx = data.targetGameX
-  const ty = data.targetGameY
-  if (sx == null || sy == null || tx == null || ty == null) return null
-  if (data.isBidirectional === true) {
-    if (nearSource) return `goes to (${tx}, ${ty})`
-    return `goes to (${sx}, ${sy})`
-  }
-  if (nearSource) return `goes to (${tx}, ${ty})`
-  return `exit - entrance at (${sx}, ${sy})`
-}
-
-function recenterViewportOnFlowPoint(
-  flowX: number,
-  flowY: number,
-  domNode: HTMLElement | null,
-  getViewport: () => { x: number; y: number; zoom: number },
-  setViewport: (vp: { x: number; y: number; zoom: number }) => void
-): void {
-  if (!domNode) return
-  const rect = domNode.getBoundingClientRect()
-  const w = Math.max(rect.width, 1)
-  const h = Math.max(rect.height, 1)
-  const vp = getViewport()
-  const z = Math.max(Number(vp.zoom) || 0.2, 0.2)
-  setViewport({ x: w / 2 - flowX * z, y: h / 2 - flowY * z, zoom: z })
-}
-
-/** Stellar Cartography wormhole edge: sky line, mono arrowhead, click recenter, hover label. */
-function WormholeEdgeOnePixel(props: EdgeProps) {
-  const { setViewport, getViewport } = useReactFlow()
-  const { sourceNode, targetNode, zoom, domNode, transform } = useStore(
-    (s) => ({
-      sourceNode: s.nodeLookup.get(props.source),
-      targetNode: s.nodeLookup.get(props.target),
-      zoom: s.transform[2],
-      domNode: s.domNode ?? null,
-      transform: s.transform,
-    }),
-    shallow
-  )
-  const scale = safeZoomScale(zoom)
-  const half = NODE_SIZE_FLOW / 2
-  const sourceX = sourceNode ? sourceNode.position.x + half : props.sourceX
-  const sourceY = sourceNode ? sourceNode.position.y + half : props.sourceY
-  const targetX = targetNode ? targetNode.position.x + half : props.targetX
-  const targetY = targetNode ? targetNode.position.y + half : props.targetY
-  const [path] = getStraightPath({
-    sourceX,
-    sourceY,
-    targetX,
-    targetY,
-  })
-  const data = props.data as WormholeEdgeData | undefined
-  const isBidirectional = data?.isBidirectional !== false
-  const setWormholeHover = useContext(WormholeHoverContext)
-  const pulseWormholeAt = useContext(WormholeRecenterPulseContext)
-  const lineReveal = useContext(WormholeLineRevealContext)
-
-  const handlePointer = useCallback(
-    (clientX: number, clientY: number): boolean => {
-      const flow = clientToFlowPosition(clientX, clientY, domNode, transform)
-      if (!flow) return false
-      const distSource = Math.hypot(flow.x - sourceX, flow.y - sourceY)
-      const distTarget = Math.hypot(flow.x - targetX, flow.y - targetY)
-      return distSource <= distTarget
-    },
-    [domNode, transform, sourceX, sourceY, targetX, targetY]
-  )
-
-  const arrowHeadLen = 8 / scale
-  const arrowAngle = Math.atan2(targetY - sourceY, targetX - sourceX)
-  const arrowA1 = arrowAngle + Math.PI - Math.PI / 7
-  const arrowA2 = arrowAngle + Math.PI + Math.PI / 7
-  const arrowHx1 = targetX + arrowHeadLen * Math.cos(arrowA1)
-  const arrowHy1 = targetY + arrowHeadLen * Math.sin(arrowA1)
-  const arrowHx2 = targetX + arrowHeadLen * Math.cos(arrowA2)
-  const arrowHy2 = targetY + arrowHeadLen * Math.sin(arrowA2)
-
-  return (
-    <>
-      <BaseEdge
-        path={path}
-        interactionWidth={12}
-        style={{
-          stroke: WORMHOLE_LINE_STROKE,
-          strokeWidth: 1 / scale,
-          opacity: WORMHOLE_EDGE_OPACITY,
-          strokeDasharray: `${5 / scale} ${4 / scale}`,
-        }}
-        onMouseMove={(e) => {
-          lineReveal.cancelClear()
-          const nearSource = handlePointer(e.clientX, e.clientY)
-          const sx = data?.sourceGameX
-          const sy = data?.sourceGameY
-          const tx = data?.targetGameX
-          const ty = data?.targetGameY
-          if (sx != null && sy != null && tx != null && ty != null) {
-            lineReveal.revealAt(nearSource ? tx : sx, nearSource ? ty : sy)
-          }
-          const label = wormholeHoverLabel(data, nearSource)
-          setWormholeHover(label != null ? [label] : null)
-        }}
-        onMouseLeave={() => {
-          setWormholeHover(null)
-          lineReveal.scheduleClear()
-        }}
-        onClick={(e) => {
-          const nearSource = handlePointer(e.clientX, e.clientY)
-          const sx = data?.sourceGameX
-          const sy = data?.sourceGameY
-          const tx = data?.targetGameX
-          const ty = data?.targetGameY
-          if (sx == null || sy == null || tx == null || ty == null) return
-          const recenterGame = nearSource ? { x: tx, y: ty } : { x: sx, y: sy }
-          recenterMapOnWormholeGameCell(
-            recenterGame.x,
-            recenterGame.y,
-            domNode,
-            getViewport,
-            setViewport,
-            pulseWormholeAt
-          )
-          e.stopPropagation()
-        }}
-      />
-      {!isBidirectional ? (
-        <polygon
-          points={`${targetX},${targetY} ${arrowHx1},${arrowHy1} ${arrowHx2},${arrowHy2}`}
-          fill={WORMHOLE_LINE_STROKE}
-          opacity={WORMHOLE_EDGE_OPACITY}
-          style={{ pointerEvents: 'none' }}
-        />
-      ) : null}
-    </>
-  )
-}
-
-/**
- * Multi-segment map edge: source → (optional game-cell waypoints) → target, 1px on screen, same style as `StraightEdgeOnePixel`.
- */
-function PolylineEdgeOnePixel(props: EdgeProps) {
-  const { sourceNode, targetNode, zoom } = useStore(
-    (s) => {
-      const nodeLookup = s.nodeLookup
-      return {
-        sourceNode: nodeLookup.get(props.source),
-        targetNode: nodeLookup.get(props.target),
-        zoom: s.transform[2],
-      }
-    },
-    shallow
-  )
-  const scale = safeZoomScale(zoom)
-  const half = NODE_SIZE_FLOW / 2
-  const data = props.data as MapEdgeData | undefined
-  const wps = data?.waypointsInGame
-  const viaFlare = data?.viaFlare === true
-  const mapEdgeStrokeStyle: CSSProperties = {
-    stroke: viaFlare ? '#facc15' : '#b1b1b7',
-    strokeWidth: 1 / scale,
-    opacity: 0.5,
-    ...(viaFlare ? { strokeDasharray: `${4 / scale} ${3 / scale}` } : {}),
-  }
-  if (!sourceNode || !targetNode) {
-    const [path] = getStraightPath({
-      sourceX: props.sourceX,
-      sourceY: props.sourceY,
-      targetX: props.targetX,
-      targetY: props.targetY,
-    })
-    return <BaseEdge path={path} style={mapEdgeStrokeStyle} />
-  }
-  const sCx = sourceNode.position.x + half
-  const sCy = sourceNode.position.y + half
-  const tCx = targetNode.position.x + half
-  const tCy = targetNode.position.y + half
-  const flowPoints = dedupeConsecutiveFlowPoints([
-    { x: sCx, y: sCy },
-    ...(Array.isArray(wps)
-      ? wps.map((g) => {
-          const { cx, cy } = flowCenterFromMapNode({ x: g.x, y: g.y })
-          return { x: cx, y: cy }
-        })
-      : []),
-    { x: tCx, y: tCy },
-  ])
-  if (flowPoints.length < 2) {
-    const [path] = getStraightPath({
-      sourceX: sCx,
-      sourceY: sCy,
-      targetX: tCx,
-      targetY: tCy,
-    })
-    return <BaseEdge path={path} style={mapEdgeStrokeStyle} />
-  }
-  const path = buildSvgPathThroughFlowPoints(flowPoints)
-  return <BaseEdge path={path} style={mapEdgeStrokeStyle} />
-}
-
-const edgeTypes = {
-  straight: StraightEdgeOnePixel,
-  polyline: PolylineEdgeOnePixel,
-  wormhole: WormholeEdgeOnePixel,
-}
-
-/** Map coordinates (px, py) are cell indices; node geometry stays fixed and centered on the map cell. */
-function toFlowNodes(nodes: CombinedMapData['nodes']): Node<MapNodeData>[] {
-  const half = NODE_SIZE_FLOW / 2
-  return nodes.map((node, i) => {
-    const x = Number(node.x)
-    const y = Number(node.y)
-    const px = Number.isFinite(x) ? x : 0
-    const py = Number.isFinite(y) ? y : 0
-    const cx = px + CELL_CENTER_OFFSET
-    const cy = gameMapYToFlowCenterY(py)
-    return {
-      id: node.id,
-      type: 'dot',
-      position: { x: cx - half, y: cy - half },
-      width: NODE_SIZE_FLOW,
-      height: NODE_SIZE_FLOW,
-      data: {
-        label: node.label,
-        ordinal: i + 1,
-        x: px,
-        y: py,
-        planet: node.planet,
-        ownerName: node.ownerName,
-      },
-    }
-  })
-}
-
-function toEdges(edges: CombinedMapData['edges']): Edge[] {
-  return edges.map((e, i) => {
-    if (e.layer === 'wormholes') {
-      return {
-        id: `e-${e.source}-${e.target}-${i}`,
-        source: e.source,
-        target: e.target,
-        sourceHandle: 's',
-        targetHandle: 't',
-        type: 'wormhole',
-        data: {
-          isBidirectional: e.isBidirectional,
-          sourceGameX: e.sourceGameX,
-          sourceGameY: e.sourceGameY,
-          targetGameX: e.targetGameX,
-          targetGameY: e.targetGameY,
-          wormholeExitOnly: e.wormholeExitOnly,
-        } satisfies WormholeEdgeData,
-      }
-    }
-    const wps = e.waypointsInGame
-    const hasPoly = Array.isArray(wps) && wps.length > 0
-    return {
-      id: `e-${e.source}-${e.target}-${i}`,
-      source: e.source,
-      target: e.target,
-      sourceHandle: 's',
-      targetHandle: 't',
-      type: hasPoly ? 'polyline' : 'straight',
-      data: {
-        viaFlare: e.viaFlare === true,
-        ...(hasPoly ? { waypointsInGame: wps } : {}),
-      } satisfies MapEdgeData,
-    }
-  })
-}
-
-/**
- * Converts client position to flow (graph) coordinates.
- * xyflow stores transform as [tx, ty, scale] where (tx, ty) is the *pane* position of the
- * flow origin, and uses flow = (pane - translation) / scale (see pointToRendererPoint usage).
- * We need the element whose (0,0) is the pane origin: the viewport or the flow container.
- */
-function clientToFlowPosition(
-  clientX: number,
-  clientY: number,
-  domNode: HTMLElement | null,
-  transform: [number, number, number] | undefined,
-  /** When supplied, avoids getBoundingClientRect (e.g. one rect per animation frame). */
-  paneRect?: Pick<DOMRect, 'left' | 'top'>
-): { x: number; y: number } | null {
-  if (!domNode || !transform) return null
-  const rect = paneRect ?? domNode.getBoundingClientRect()
-  const [tx, ty, scale] = transform
-  if (typeof scale !== 'number' || !Number.isFinite(scale) || scale <= 0) return null
-  const paneX = clientX - rect.left
-  const paneY = clientY - rect.top
-  const x = (paneX - tx) / scale
-  const y = (paneY - ty) / scale
-  if (!Number.isFinite(x) || !Number.isFinite(y)) return null
-  return { x, y }
-}
 
 /**
  * Computes initial viewport so the bounding rectangle of all node centers is
@@ -845,262 +332,6 @@ function NormalWarpWellOutlinesOverlay({ mapNodes }: { mapNodes: CombinedMapData
   )
 }
 
-function StellarCartographyRadialGradientDef({
-  gradient,
-  variant,
-}: {
-  gradient: StellarCartographyOverlayRadialGradient
-  variant: 'core' | 'band'
-}) {
-  const innerStop = `${gradient.innerOffset * 100}%`
-  if (variant === 'core') {
-    return (
-      <radialGradient id={gradient.id} cx="50%" cy="50%" r="50%">
-        <stop offset="0%" stopColor={gradient.color} stopOpacity={gradient.peakOpacity} />
-        <stop offset={innerStop} stopColor={gradient.color} stopOpacity={gradient.peakOpacity} />
-        <stop offset="100%" stopColor={gradient.color} stopOpacity={gradient.edgeOpacity} />
-      </radialGradient>
-    )
-  }
-  return (
-    <radialGradient id={gradient.id} cx="50%" cy="50%" r="50%">
-      <stop offset="0%" stopColor={gradient.color} stopOpacity={0} />
-      <stop offset={innerStop} stopColor={gradient.color} stopOpacity={0} />
-      <stop offset={innerStop} stopColor={gradient.color} stopOpacity={gradient.peakOpacity} />
-      <stop offset="100%" stopColor={gradient.color} stopOpacity={gradient.edgeOpacity} />
-    </radialGradient>
-  )
-}
-
-function StellarCartographyOverlay({
-  overlayCircles,
-  wormholeEndpoints,
-  wormholeEndpointHoverByCell,
-  wormholeRecenterPulseTarget,
-  blockedByPlanetHover,
-  nuIonStorms,
-}: {
-  overlayCircles: CombinedMapData['overlayCircles']
-  wormholeEndpoints: { x: number; y: number }[]
-  wormholeEndpointHoverByCell: Map<string, WormholeEndpointHoverInfo>
-  wormholeRecenterPulseTarget: WormholeRecenterPulseTarget | null
-  blockedByPlanetHover: boolean
-  nuIonStorms?: boolean
-}) {
-  const { setViewport, getViewport } = useReactFlow()
-  const setWormholeHover = useContext(WormholeHoverContext)
-  const pulseWormholeAt = useContext(WormholeRecenterPulseContext)
-  const lineReveal = useContext(WormholeLineRevealContext)
-  const domNode = useStore((s) => s.domNode ?? null)
-  const transform = useStore((s) => s.transform)
-  const [size, setSize] = useState({ width: 0, height: 0 })
-
-  useEffect(() => {
-    if (!domNode) return
-    let raf = 0
-    const ro = new ResizeObserver((entries) => {
-      const { width, height } = entries[0]?.contentRect ?? { width: 0, height: 0 }
-      cancelAnimationFrame(raf)
-      raf = requestAnimationFrame(() => setSize({ width, height }))
-    })
-    ro.observe(domNode)
-    return () => {
-      cancelAnimationFrame(raf)
-      ro.disconnect()
-    }
-  }, [domNode])
-
-  if (!transform || size.width <= 0 || size.height <= 0) return null
-  if (overlayCircles.length === 0 && wormholeEndpoints.length === 0) return null
-
-  const [tx, ty, rawScale] = transform
-  const scale = safeZoomScale(rawScale)
-  const { width, height } = size
-  const shapes = buildStellarCartographyOverlayPaneShapes(overlayCircles, wormholeEndpoints, {
-    width,
-    height,
-    tx,
-    ty,
-    scale,
-  }, { cloudyIonStorms: nuIonStorms ?? true })
-
-  if (
-    shapes.circles.length === 0 &&
-    shapes.annuli.length === 0 &&
-    shapes.nebulaClouds.length === 0 &&
-    shapes.ionStormClouds.length === 0 &&
-    shapes.debrisDiskBorders.length === 0 &&
-    shapes.arrows.length === 0 &&
-    shapes.wormholeMarkers.length === 0
-  ) {
-    return null
-  }
-
-  return (
-    <div className="pointer-events-none absolute inset-0 z-[5]" aria-hidden>
-      <svg className="h-full w-full" viewBox={`0 0 ${width} ${height}`} preserveAspectRatio="none">
-        {shapes.nebulaClouds.map((shape) => (
-          <RasterFieldOverlay key={shape.key} {...nebulaCloudPaneShapeToRasterField(shape)} />
-        ))}
-        {shapes.ionStormClouds.map((shape) => (
-          <RasterFieldOverlay key={shape.key} {...ionStormCloudPaneShapeToRasterField(shape)} />
-        ))}
-        {shapes.circles.map(({ key, cx, cy, r, fill, stroke, strokeWidth, fillGradient }) => (
-          <g key={key}>
-            {fillGradient != null && (
-              <defs>
-                <StellarCartographyRadialGradientDef gradient={fillGradient} variant="core" />
-              </defs>
-            )}
-            <circle
-              cx={cx}
-              cy={cy}
-              r={r}
-              fill={fillGradient != null ? `url(#${fillGradient.id})` : fill}
-              stroke={stroke}
-              strokeWidth={strokeWidth}
-            />
-          </g>
-        ))}
-        {shapes.annuli.map(
-          ({
-            key,
-            cx,
-            cy,
-            coreR,
-            bandR,
-            coreFill,
-            coreStroke,
-            coreGradient,
-            bandFill,
-            bandStroke,
-            strokeWidth,
-            bandGradient,
-          }) => (
-            <g key={key}>
-              {(bandGradient != null || coreGradient != null) && (
-                <defs>
-                  {bandGradient != null && (
-                    <StellarCartographyRadialGradientDef gradient={bandGradient} variant="band" />
-                  )}
-                  {coreGradient != null && (
-                    <StellarCartographyRadialGradientDef gradient={coreGradient} variant="core" />
-                  )}
-                </defs>
-              )}
-              <circle
-                cx={cx}
-                cy={cy}
-                r={bandR}
-                fill={bandGradient != null ? `url(#${bandGradient.id})` : bandFill}
-                stroke={bandStroke}
-                strokeWidth={strokeWidth}
-              />
-              <circle
-                cx={cx}
-                cy={cy}
-                r={coreR}
-                fill={
-                  coreGradient != null ? `url(#${coreGradient.id})` : coreFill
-                }
-                stroke={coreStroke ?? 'none'}
-                strokeWidth={coreStroke != null ? strokeWidth : 0}
-              />
-            </g>
-          )
-        )}
-        {shapes.debrisDiskBorders.map(({ key, cx, cy, r, fill, stroke, strokeWidth }) => (
-          <circle
-            key={key}
-            cx={cx}
-            cy={cy}
-            r={r}
-            fill={fill}
-            stroke={stroke}
-            strokeWidth={strokeWidth}
-          />
-        ))}
-        {shapes.arrows.map(({ key, x1, y1, x2, y2, stroke, strokeWidth }) => {
-          const angle = Math.atan2(y2 - y1, x2 - x1)
-          const headLen = 6
-          const a1 = angle + Math.PI - Math.PI / 7
-          const a2 = angle + Math.PI + Math.PI / 7
-          const hx1 = x2 + headLen * Math.cos(a1)
-          const hy1 = y2 + headLen * Math.sin(a1)
-          const hx2 = x2 + headLen * Math.cos(a2)
-          const hy2 = y2 + headLen * Math.sin(a2)
-          return (
-            <g key={key}>
-              <line
-                x1={x1}
-                y1={y1}
-                x2={x2}
-                y2={y2}
-                stroke={stroke}
-                strokeWidth={strokeWidth}
-              />
-              <polygon points={`${x2},${y2} ${hx1},${hy1} ${hx2},${hy2}`} fill={stroke} />
-            </g>
-          )
-        })}
-      </svg>
-      <div className="absolute inset-0" aria-hidden>
-        {shapes.wormholeMarkers.map(({ key, cx, cy, diameterPx, mapX, mapY }) => {
-          const half = diameterPx / 2
-          const hoverInfo = wormholeEndpointHoverByCell.get(wormholeMapCellKey(mapX, mapY))
-          const recenterGame =
-            hoverInfo != null ? wormholeEndpointRecenterGameCoords(hoverInfo) : null
-          const isPulseTarget =
-            wormholeRecenterPulseTarget != null &&
-            wormholeRecenterPulseTarget.mapX === mapX &&
-            wormholeRecenterPulseTarget.mapY === mapY
-          return (
-            <div
-              key={key}
-              className={`absolute pointer-events-auto${recenterGame != null ? ' cursor-pointer' : ''}`}
-              style={{
-                left: cx - half,
-                top: cy - half,
-                width: diameterPx,
-                height: diameterPx,
-              }}
-              onMouseEnter={() => {
-                lineReveal.cancelClear()
-                lineReveal.revealAt(mapX, mapY)
-                if (blockedByPlanetHover || hoverInfo == null) return
-                setWormholeHover(formatWormholeEndpointHoverLines(hoverInfo))
-              }}
-              onMouseLeave={() => {
-                setWormholeHover(null)
-                lineReveal.scheduleClear()
-              }}
-              onClick={(e) => {
-                if (recenterGame == null) return
-                recenterMapOnWormholeGameCell(
-                  recenterGame.x,
-                  recenterGame.y,
-                  domNode,
-                  getViewport,
-                  setViewport,
-                  pulseWormholeAt
-                )
-                e.stopPropagation()
-              }}
-            >
-              <div
-                key={isPulseTarget ? `pulse-${wormholeRecenterPulseTarget.token}` : 'idle'}
-                className={`h-full w-full${isPulseTarget ? ' wormhole-recenter-pulse' : ''}`}
-              >
-                <WormholeEndpointIconMark />
-              </div>
-            </div>
-          )
-        })}
-      </div>
-    </div>
-  )
-}
-
 /**
  * in size regardless of zoom. Uses same flow->pane conversion as the grid.
  */
@@ -1126,17 +357,6 @@ function buildLabelSourceByNodeId(nodes: CombinedMapData['nodes']): Map<string, 
     m.set(n.id, payload)
   }
   return m
-}
-
-/** Flow-space center of the planet dot; must match `toFlowNodes` + half offset. */
-function flowCenterFromMapNode(mapNode: { x: number; y: number }): { cx: number; cy: number } {
-  const x = Number(mapNode.x)
-  const y = Number(mapNode.y)
-  const px = Number.isFinite(x) ? x : 0
-  const py = Number.isFinite(y) ? y : 0
-  const cx = px + CELL_CENTER_OFFSET
-  const cy = gameMapYToFlowCenterY(py)
-  return { cx, cy }
 }
 
 function FixedSizeDotsOverlay({
@@ -1728,7 +948,7 @@ export function MapGraph({
             <ViewportZoomSync onMapZoomChange={onMapZoomChange} />
             <SliderZoomControl onMapZoomChange={onMapZoomChange} onSetZoomReady={onSetZoomReady} />
             <CoordinateGridOverlay />
-            <StellarCartographyOverlay
+            <StellarCartographyOverlayPane
               overlayCircles={data.overlayCircles}
               wormholeEndpoints={wormholeEndpoints}
               wormholeEndpointHoverByCell={wormholeEndpointHoverByCell}
