+++ packages/frontend/src/lib/warpWellOverlay.test.ts
@@ -57,7 +57,9 @@ describe('warp well overlay regression', () => {
     const normalized = normalizeMapDataResponse(raw)
     expect(normalized.nodes[0].normalWellCells).toHaveLength(29)
 
-    const combined = combineMapData(['base-map'], [{ data: normalized }], null)
+    const combined = combineMapData(['base-map'], [{ data: normalized }], {
+      liveConnectionsParams: null,
+    })
     expect(combined.nodes[0].normalWellCells).toHaveLength(29)
   })
